requirements.txt
slackclient>=2.0.0
beautifulsoup4
holidays